rename index modules in scripts
